backend/requirements.txt
djangorestframework==3.15.2
Django==5.0.8
django-rest-swagger==2.2.0
django-cors-headers==4.4.0
drf-yasg==1.21.7
psycopg2==2.9.9
PyJWT==2.3.0
